Stupid that IdClass maps based on field name apparently + postgis bump
